==> 1_colecting_community_data_from_miningpoolstats.ipynb <==
from selenium import webdriver
from selenium.webdriver.chrome.options import Options as ChromeOptions

import pandas as pd
import time
pd.options.display.max_rows = None

def _driver():
    driver_path = './chromedriver.exe'
    driver = webdriver.Chrome(executable_path=driver_path)
    return driver

# ABRE O NEVEGADOR
driver = _driver()
main_url = 'https://miningpoolstats.stream'
driver.get(main_url)

#PEGA A PRIMEIRA PAGINA (CLICAR MANUALMENTE EM SHOW ALL RESULTS)
htmlPage = bs4.BeautifulSoup(driver.page_source, features='html.parser')
resultados = htmlPage.find_all('tr',{'role':'row'})
resultados.pop(0) #exclui o primeiro resultado que é as colunas

#FAZ PARSE NOS DADOS E SALVA UM DATAFRAME
data = []
for item in resultados:
    row_data = {'link':'', 'coin_name':'', 'sigla':'', 'algorithm':'', 'preco_dolar':'', 'pools':'',
           'total_hashare':'', 'network_hashate': ''}
    
    link = item.find('a')
    if link != None:  row_data['link'] = main_url + link.get('href')
    
    coin_name = item.find('b')
    if coin_name != None:  row_data['coin_name'] = coin_name.text
    
    sigla = item.find('small')
    if sigla != None:  row_data['sigla'] = sigla.text
    
    algorithm = item.find('div',{'style':'text-align:center;padding-right:25px;padding-top:5px;'})
    if algorithm != None:  row_data['algorithm'] = algorithm.text
    
    preco_dolar = item.find('td',{'class':'show1050'})
    if preco_dolar != None:  row_data['preco_dolar'] = preco_dolar.text
    
    pools = item.find('td',{'class':'show1150'})
    if pools != None:  row_data['pools'] = pools.text
    
    total_hashare = item.find('td',{'class':'show1250'})
    if total_hashare != None:  row_data['total_hashare'] = total_hashare.text

    network_hashate = item.find('td',{'class':'show1350'})
    if network_hashate != None:  row_data['network_hashate'] = network_hashate.text
    
    data.append(row_data)

data = pd.DataFrame(data)

def getCoinLinks(position_df):
    coinPage = bs4.BeautifulSoup(req.text)
    informations = coinPage.find('div', {'style':'padding-left:0px;padding-right:0px;'})
    
    #PEGA OS DADOS QUE VARIAM E HORA PODEM NAO APARECER
    for item in informations.find_all('tr'):
        try:
            found = item.find('a').get('href')
            data2.loc[position_df, item.find_all('td')[0].text] = found
        except:
            pass

data2 = data.copy()
for pos, link in data['link'].T.items():
    #faz o request para cada pagina de criptomoeda dentro do miningpool para pegar os links de referencia para mais dados
    req = requests.get(link)
    if req.status_code == 200:
        getCoinLinks(pos)
        print('{} coin processed {}'.format(pos, data.loc[pos, 'coin_name']))
    else:
        print('{} problem with coin {}'.format(pos, data.loc[pos, 'coin_name']))
    
    time.sleep(3)

data2.columns

# data2['Calculator ']
data2.drop(['Calculator ', 'Info ', '\n\n \n'], axis=1, inplace=True)

data2.to_csv('community_data_720Coins_dataFrom_miningpoolstarts.csv', index=None)
