==> 2_data_from_each_website.ipynb <==
import requests
import pandas as pd
import time, bs4
pd.options.display.max_rows = None

data_old = pd.read_csv('data/community_data_720Coins_dataFrom_miningpoolstarts.csv')

data.columns
extra_columns = ['Reddit ', 'Twitter ', 'Forum ', 'Chat ', 'Telegram ']
data.drop(extra_columns, axis=1, inplace=True)

data.head()

def getTelegramInfo(telegram_link):
    try:
        req = requests.get(telegram_link)
        if req.status_code == 200:
            htmlPage = bs4.BeautifulSoup(req.content, features='html.parser')
            return htmlPage.find('div', {'class':'tgme_page_extra'}).text
        else:
            return 'Error'
    except:
        return 'Error'

# getTelegramInfo('https://t.me/Conflux_English')

def getWebsiteHtml(website_link):
    try:
        req = requests.get(website_link)
        if req.status_code == 200:
            htmlPage = bs4.BeautifulSoup(req.content, features='html.parser')

            #SAVE HTML FILE
            path_ = 'html_sites_cryptocurrencies/'+ str(website_link.replace('http://', '').replace('https://', '')) + '.txt' 
            with open(path_, 'w') as filee:
                filee.write(str(htmlPage.encode('utf8')))
            return htmlPage
        else:
            return 'Error'
    except:
        return 'Error'

def getSocialMedia(position_dataframe, html_file):
    for link in html_file.find_all('a'):
        if link.get('href') != None and 'facebook' in link.get('href'): data.loc[position_dataframe, 'facebook'] = link.get('href')

        if link.get('href') != None and 't.me' in link.get('href'): data.loc[position_dataframe, 'telegram'] = link.get('href')

        if link.get('href') != None and 'instagram' in link.get('href'): data.loc[position_dataframe, 'instagram'] = link.get('href')

        if link.get('href') != None and 'discord' in link.get('href'): data.loc[position_dataframe, 'discord'] = link.get('href')

        if link.get('href') != None and 'twitter' in link.get('href'): data.loc[position_dataframe, 'twitter'] = link.get('href')

        if link.get('href') != None and 'youtube' in link.get('href'): data.loc[position_dataframe, 'youtube'] = link.get('href')

        if link.get('href') != None and 'bitcointalk' in link.get('href'): data.loc[position_dataframe, 'bitcointalk'] = link.get('href')

        if link.get('href') != None and 'github' in link.get('href'): data.loc[position_dataframe, 'github'] = link.get('href')

        if link.get('href') != None and 'reddit' in link.get('href'): data.loc[position_dataframe, 'reddit'] = link.get('href')

for pos, link in data[data['Website '].isnull() == False]['Website '].T.items():
    #SE DER ALGUM ERRO NO REQUESTS IRA SALVAR O CODIGO DE RESPOSTA DO SITE
    response_ = getWebsiteHtml(link)
    if response_ == 'Error': 
        data.loc[pos, 'site_error'] = response_
    else:
        getSocialMedia(pos, response_)
        
    print('{} website processed {}'.format(pos, data.loc[pos, 'coin_name']))

## Segundo passo para dados que nao foram capturados com request

from selenium import webdriver
from selenium.webdriver.chrome.options import Options as ChromeOptions

def _driver():
    driver_path = './chromedriver.exe'
    driver = webdriver.Chrome(executable_path=driver_path)
    return driver

driver = _driver()

def getWebsiteHtmlBySelenium(website_link):
    try:
        driver.get(website_link)
        htmlPage = bs4.BeautifulSoup(driver.page_source, features='html.parser')
        #SAVE HTML FILE
        path_ = 'html_sites_cryptocurrencies/'+ str(website_link.replace('http://', '').replace('https://', '')) + '.txt' 
        with open(path_, 'w') as filee:
            filee.write(str(htmlPage.encode('utf8')))
        return htmlPage
    except:
        return 'Error'

#PEGA OS SITES QUE NAO CONSEGUIU NENHUM DADO PORQUE PODE TER DADO PROBLEMA COM REQUEST
next_step = data[(data.github.isnull() == True) & (data.twitter.isnull() == True) & (data.youtube.isnull() == True) & (data.discord.isnull() == True) & (data.telegram.isnull() == True) & (data.reddit.isnull() == True) & (data.facebook.isnull() == True) & (data.bitcointalk.isnull() == True) & (data.instagram.isnull() == True)]

for pos, link in next_step['Website '].T.items():
    #SE DER ALGUM ERRO NO REQUESTS IRA SALVAR O CODIGO DE RESPOSTA DO SITE
    response_ = getWebsiteHtmlBySelenium(link)
    if response_ == 'Error': 
        data.loc[pos, 'site_error'] = response_
    else:
        getSocialMedia(pos, response_)
        
    print('{} website processed {}'.format(pos, data.loc[pos, 'coin_name']))

#PARA COINS QUE NAO ENCONTROU O TELEGRAM NO SITE ANTERIOR POREM ENCONTROU LA NO MINIGPOOL
for pos, item in data[data['telegram'].isnull() ==True]['telegram'].T.iteritems():
    if pd.isna(data_old.loc[pos, 'Telegram ']):
        pass
    else:
        data.loc[pos, 'telegram'] = data_old.loc[pos, 'Telegram ']

data.to_csv('data/SocialMedia_community_data_720Coins_dataFrom_eachWebsite.csv', index=None)
