==> 3_social_media_DETAILS.ipynb <==
## DADOS DO TELEGRAM

data = pd.read_csv('data/SocialMedia_community_data_720Coins_dataFrom_eachWebsite.csv')

from datetime import datetime

for pos, link in data[data['telegram'].isnull() == False]['telegram'].T.items():
    if pos < 87:
        pass
    else:
        telegram_info = getTelegramInfo(link)
        data.loc[pos, 'telegram_info'] = telegram_info

        print('{} website processed {} has in telegram {}'.format(pos, data.loc[pos, 'coin_name'], telegram_info))
    #     break
    
# current date and time
now = datetime.now()
timestamp = datetime.timestamp(now)

data['last_telegram_update_timestamp'] = timestamp

data[data['telegram_info'].isnull() ==False]

data.to_csv('data/SocialMedia_community_data_720Coins_dataFrom_eachWebsite.csv', index=None)
